confiscated_ms/requirements.txt
Django==2.2.1
djangorestframework==3.9.4
django-filter
mysqlclient
requests